Ported to require.js + changed tabs to spaces
I'm using this in a project that runs client-side as well, and I am
using require.js for it.  I've wrapped all the files in define() methods
and made a few other minor changes.  I also believe in spaces for
indentation instead of tabs ;)

This hasn't been tested since I've made the changes so there may be some
bugs.